feat: type Event.distance as float for non-integer distances (#29)
The E1/F1 distance field (6 chars) can contain non-integer values — e.g.
"2.4" for 2.4 miles in open-water meets. safe_cast(int, "2.4") raised
ValueError and fell back to int() = 0, permanently losing the value.

Type Event.distance as float and parse it with safe_cast(float, ...) so
these distances survive. Integer distances now come through as floats
(100 -> 100.0); this is an API-compatible data change, so callers reading
distance as an int should coerce. Blank/non-numeric fields still yield 0.0.

+++ hytek_parser/hy3/line_parsers/f_relay_parsers.py
@@ -30,7 +30,7 @@ def f1_parser(
     # Get event info
     event_gender = select_from_enum(Gender, extract(line, 14, 1))
     event_gender_age = select_from_enum(GenderAge, extract(line, 15, 1))
-    distance = safe_cast(int, extract(line, 16, 6))
+    distance = safe_cast(float, extract(line, 16, 6))
     stroke = select_from_enum(Stroke, extract(line, 22, 1))
 
     # Have to set these now since swimmer ages are not available yet
